requirements.txt
django
django-storages[azure]
whitenoise
numpy
gremlinpython==3.7.0
git+https://github.com/azure-samples/ms-identity-python-utilities@main
pyyaml

# Some extra libraries used in demo (not needed in the application)
# pandas